# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: upsertLead
      tags: [leads]
      summary: Create or update a storefront lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "200":
          description: Lead saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerLead"
  /quotes:
    get:
      operationId: getSavedQuote
      tags: [quotes]
      summary: Get a saved studio quotation by signed public token
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 512
      responses:
        "200":
          description: Saved quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerLead"
        "404":
          description: Quote not found
  /quotes/notify:
    post:
      operationId: notifySavedQuote
      tags: [quotes]
      summary: Send a saved quotation to the configured sales channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyQuoteInput"
      responses:
        "200":
          description: Quote notification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteNotificationResponse"
        "404":
          description: Quote not found
  /leads/sketch:
    post:
      operationId: submitSketchLead
      tags: [leads]
      summary: Save a hand-sketch lead with its uploaded image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, metadata]
              properties:
                file: { type: string, description: Uploaded image file supplied as multipart content }
                metadata: { type: string, description: JSON-encoded LeadInput metadata }
      responses:
        "201":
          description: Sketch lead saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SketchLeadResponse"
  /leads/payment-slip:
    post:
      operationId: submitPaymentSlip
      tags: [leads]
      summary: Upload a bank transfer slip and verify it against the quote total via SlipOK
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, token]
              properties:
                file: { type: string, description: Uploaded slip image supplied as multipart content }
                token: { type: string, description: Public quote access token, minLength: 32, maxLength: 512 }
                kind:
                  type: string
                  enum: [deposit, final]
      responses:
        "201":
          description: Payment slip saved and verification result recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSlip"
        "404":
          description: Quote not found
...
components:
...
  schemas:
...
    LeadInput:
      type: object
      required: [leadKey, status, source, productSkus]
      properties:
        leadKey: { type: string, minLength: 12, maxLength: 120 }
        status:
          type: string
          enum: [new_lead, selecting, quote_requested, closed]
        source: { type: string, maxLength: 64 }
        name: { type: [string, "null"], maxLength: 160 }
        company: { type: [string, "null"], maxLength: 200 }
        phone: { type: [string, "null"], maxLength: 64 }
        lineContact: { type: [string, "null"], maxLength: 120 }
        email: { type: [string, "null"], maxLength: 240 }
        project: { type: [string, "null"], maxLength: 240 }
        address: { type: [string, "null"] }
        site: { type: [string, "null"], maxLength: 240 }
        purchasingDepartment: { type: [string, "null"], maxLength: 160 }
        notes: { type: [string, "null"] }
        taxName: { type: [string, "null"], maxLength: 240 }
        taxId: { type: [string, "null"], pattern: "^$|^[0-9]{13}$" }
        taxBranch: { type: [string, "null"], maxLength: 120 }
        taxAddress: { type: [string, "null"] }
        preferredContact:
          type: [string, "null"]
          enum: [line, phone, email, null]
        customerRole:
          type: [string, "null"]
          enum: [homeowner, architect-interior, contractor, null]
        propertyType:
          type: [string, "null"]
          enum: [house-townhome, condo, commercial, null]
        condoFloor: { type: [string, "null"], maxLength: 32 }
        expectedInstallationDate: { type: [string, "null"], format: date }
        productSkus:
          type: array
          maxItems: 20
          items: { type: string, maxLength: 32 }
        quoteNumber: { type: [string, "null"], maxLength: 64 }
        orderMode:
          type: string
          enum: [quick-purchase, studio, sketch]
        studioData:
          type: [object, "null"]
          additionalProperties: true
        sketchUrl: { type: [string, "null"], maxLength: 2000 }
...
    CustomerLead:
      type: object
      required: [id, leadKey, status, source, productSkus, createdAt, updatedAt]
      properties:
        id: { type: integer }
        leadKey: { type: string }
        status:
          type: string
          enum: [new_lead, selecting, quote_requested, closed]
        source: { type: string }
        name: { type: [string, "null"] }
        company: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        lineContact: { type: [string, "null"] }
        email: { type: [string, "null"] }
        project: { type: [string, "null"] }
        address: { type: [string, "null"] }
        site: { type: [string, "null"] }
        purchasingDepartment: { type: [string, "null"] }
        notes: { type: [string, "null"] }
        taxName: { type: [string, "null"] }
        taxId: { type: [string, "null"] }
        taxBranch: { type: [string, "null"] }
        taxAddress: { type: [string, "null"] }
        preferredContact: { type: [string, "null"] }
        customerRole: { type: [string, "null"] }
        propertyType: { type: [string, "null"] }
        condoFloor: { type: [string, "null"] }
        expectedInstallationDate: { type: [string, "null"], format: date }
        productSkus:
          type: array
          items: { type: string }
        quoteNumber: { type: [string, "null"] }
        publicQuoteToken: { type: [string, "null"] }
        orderMode:
          type: string
          enum: [quick-purchase, studio, sketch]
        studioData:
          type: [object, "null"]
          additionalProperties: true
        sketchUrl: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NotifyQuoteInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 32, maxLength: 512 }
    QuoteNotificationResponse:
      type: object
      required: [notificationStatus, message]
      properties:
        notificationStatus:
          type: string
          enum: [notified, saved_not_notified]
        message: { type: string }
    SketchLeadResponse:
      type: object
      required: [lead, notificationStatus]
      properties:
        lead:
          $ref: "#/components/schemas/CustomerLead"
        notificationStatus:
          type: string
          enum: [notified, saved_not_notified]
        message: { type: string }
    PaymentSlip:
      type: object
      required: [id, leadId, kind, status, slipImageUrl, reviewedByAdmin, createdAt, updatedAt]
      properties:
        id: { type: integer }
        leadId: { type: integer }
        kind:
          type: string
          enum: [deposit, final]
        status:
          type: string
          enum: [pending, verified, rejected]
        slipImageUrl: { type: string }
        claimedAmountThb: { type: [integer, "null"] }
        verifiedAmountThb: { type: [integer, "null"] }
        senderName: { type: [string, "null"] }
        transRef: { type: [string, "null"] }
        slipokErrorCode: { type: [string, "null"] }
        reviewedByAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...